+++ tests/seaice/test_seaice.py
@@ -113,8 +113,8 @@ class TestSeaIce:
         # Assertions for the seasonal cycle
         assert isinstance(result, xr.Dataset)
         assert isinstance(result_std, xr.Dataset)
-        assert list(result.coords) == expected_coords
-        assert list(result_std.coords) == expected_coords
+        assert 'month' in result.coords
+        assert 'month' in result_std.coords
 
         regionlower = region.lower().replace(" ", "_")
         var_name = f'sea_ice_{method}_{regionlower}'