+++ CHANGELOG.md
@@ -14,6 +14,7 @@ AQUA core complete list:
 AQUA diagnostics complete list:
 - LatLonProfiles: tests implementation (#2339)
 - Ocean Drift: timeseries plotting function with update in cli (#2322)
+- Implement `fldstat` methods in Seaice diagnostics (#2297)
 - Gregory: more detailed description and plot labels (#2306)
 
 ## [v0.18.0]